make clean be silent in happy path.

+++ clean.sh
@@ -1,7 +1,25 @@
 #!/bin/bash
 
-rm -Rf bin
-rm -Rf build
-rm hostyarg/hostyarg
-rm hostyarg/test.uf2
-rm -Rf cyarg/build
+if [ -d bin ]
+then
+    rm -R bin
+fi
+if [ -d build ]
+then
+    rm -R build
+fi
+
+if [ -e hostyarg/hostyarg ]
+then
+    rm hostyarg/hostyarg
+fi
+
+if [ -e hostyarg/test.uf2 ]
+then
+    rm hostyarg/test.uf2
+fi
+
+if [ -d cyarg/build ]
+then
+    rm -R cyarg/build
+fi
